More changes to MDT CSV dumper for ACTS standalone interoperability
More changes to MDT CSV dumper for ACTS standalone interoperability

+++ MuonSpectrometer/MuonPhaseII/MuonCnv/MuonHitCsvDump/src/CsvMdtDriftCircleDumperMuonCnv.h
@@ -32,6 +32,6 @@ class CsvMdtDriftCircleDumperMuonCnv: public AthAlgorithm {
     ServiceHandle<Muon::IMuonIdHelperSvc> m_idHelperSvc{
         this, "MuonIdHelperSvc", "Muon::MuonIdHelperSvc/MuonIdHelperSvc"};
 
-   int m_event = 0;
+   size_t m_event = 0;
 
 };
